+++ src/aks-preview/azext_aks_preview/_params.py
@@ -40,6 +40,7 @@ from azext_aks_preview._validators import (
     validate_nat_gateway_v2_params_for_update,
     validate_outbound_type_sku,
     validate_outbound_type_sku_for_update,
+    validate_action_group_id,
 )
 from azext_aks_preview._client_factory import CUSTOM_MGMT_AKS_PREVIEW
 from azext_aks_preview._completers import (
@@ -2552,6 +2553,11 @@ def load_arguments(self, _):
                  'Example: \'[{"type":"Standard","vnetSubnetId":"/subscriptions/.../subnets/mysubnet"}]\'',
             is_preview=True,
         )
+        c.argument(
+            "enable_managed_dranet",
+            action="store_true",
+            is_preview=True,
+        )
         # prepared image specification
         c.argument(
             'prepared_image_specification_id',
@@ -2605,6 +2611,11 @@ def load_arguments(self, _):
         )
         c.argument(
             "asg_ids", validator=validate_application_security_groups, is_preview=True
+        )
+        c.argument(
+            "enable_managed_dranet",
+            action="store_true",
+            is_preview=True,
         )
         c.argument(
             "enable_artifact_streaming",
@@ -3898,6 +3909,36 @@ def load_arguments(self, _):
                      "label selector with 'matchLabels' (an array of \"key=value\" strings) and/or "
                      "'matchExpressions'. Maximum 100 entries.",
             )
+
+    # AKS alert configuration commands
+    with self.argument_context("aks alert-config") as c:
+        c.argument("cluster_name", help="The cluster name.")
+        c.argument(
+            "aks_custom_headers",
+            help="Send custom headers. When specified, format should be Key1=Value1,Key2=Value2.",
+        )
+
+    for scope in ['aks alert-config add',
+                  'aks alert-config update',
+                  'aks alert-config delete',
+                  'aks alert-config show']:
+        with self.argument_context(scope) as c:
+            c.argument('name', options_list=['--name', '-n'], required=True,
+                       help='Name of the alert configuration.')
+
+    for scope in ['aks alert-config add',
+                  'aks alert-config update']:
+        with self.argument_context(scope) as c:
+            c.argument('mode', arg_type=get_enum_type(['Disabled', 'Managed']),
+                       help='How AKS manages alerts for the cluster.')
+            c.argument('action_group_id', options_list=['--action-group-id'],
+                       validator=validate_action_group_id,
+                       help='Resource ID of the Azure Monitor action group to send '
+                            'notifications to. Pass an empty string to clear it.')
+
+    with self.argument_context('aks alert-config add') as c:
+        c.argument('mode', arg_type=get_enum_type(['Disabled', 'Managed']), required=True,
+                   help='How AKS manages alerts for the cluster.')
 
     # aks list-vm-skus command
     with self.argument_context("aks list-vm-skus") as c: