+++ src/containerapp/HISTORY.rst
@@ -4,6 +4,9 @@ Release History
 ===============
 upcoming
 ++++++
+
+1.3.0b5
+++++++
 * 'az containerapp env storage set': Support Azure Files authentication with an Azure Key Vault secret or managed identity
 
 1.3.0b4